+++ src/client.ts
@@ -29,6 +29,7 @@ import {
 
 import { Position, Order, Side } from "./types";
 import { EventType } from "./events";
+import * as constants from "./constants";
 
 export class Client {
   /**